feat: unified settings sidebar navigation & section reorganization (#1324)
