+++ frontend/src/pages/book_detail/body.rs
@@ -295,7 +295,7 @@ fn SuggestionsConnectCard(is_admin: bool) -> Element {
             }
             if is_admin {
                 Link {
-                    to: Route::Settings {},
+                    to: Route::Settings { section: Some("metadata".into()) },
                     class: "btn primary sm",
                     "data-testid": "suggestions-connect-link",
                     "Add Hardcover API key \u{2192}"
