+++ ui_tests/playwright/tests/flows/logs.spec.ts
@@ -3,10 +3,12 @@ import { expect, test } from "../fixtures/test";
 import { expectMutation } from "../utils/api";
 import { expectNavVisible, gotoReady } from "../utils/nav";
 
-// The admin log viewer (`/logs`, F5.2). The layout test runs against the real
-// on-disk logs the server is already writing (proving AC1 end-to-end); the
-// filter / error / pagination tests mock `/api/rpc/logs` after the initial
-// load so the rows under assertion are deterministic.
+// The admin log viewer — now the Logs section of Settings (#1324); the legacy
+// `/logs` route redirects here. The layout test runs against the real on-disk
+// logs the server is already writing (proving AC1 end-to-end); the filter /
+// error / pagination tests mock `/api/rpc/logs` after the initial load so the
+// rows under assertion are deterministic.
+const LOGS = "/settings?section=logs";
 
 const levelFilter = (page: Page) => page.getByLabel("Level");
 const moduleFilter = (page: Page) => page.getByLabel("Module");
@@ -59,20 +61,21 @@ test("renders the log viewer layout from the on-disk logs", async ({
   await expectMutation(
     page,
     { method: "POST", url: "/api/rpc/logs", expectedStatus: 200 },
-    async () => gotoReady(page, "/logs"),
+    async () => gotoReady(page, LOGS),
   );
 
   await expect(
-    page.getByRole("heading", { level: 1, name: "Server logs" }),
+    page.getByRole("heading", { level: 2, name: "Server logs" }),
   ).toBeVisible();
+  await expect(page.getByTestId("settings-nav-logs")).toHaveAttribute(
+    "aria-current",
+    "page",
+  );
   await expect(levelFilter(page)).toBeVisible();
   await expect(moduleFilter(page)).toBeVisible();
   await expect(page.getByLabel("From")).toBeVisible();
   await expect(page.getByLabel("To")).toBeVisible();
   await expect(applyButton(page)).toBeVisible();
-  await expect(
-    page.getByRole("link", { name: "Back to settings" }),
-  ).toBeVisible();
   // The server is continuously logging its own request handling, so the
   // real on-disk logs are non-empty — the table renders actual records.
   await expect(page.getByTestId("logs-table")).toBeVisible();
@@ -83,7 +86,7 @@ test("renders the log viewer layout from the on-disk logs", async ({
 test("filtering by level narrows the results and sends the level in the query", async ({
   page,
 }) => {
-  await gotoReady(page, "/logs");
+  await gotoReady(page, LOGS);
 
   // Mock only the post-Apply request so the rendered rows are deterministic.
   await mockLogs(
@@ -115,7 +118,7 @@ test("filtering by level narrows the results and sends the level in the query", 
 test("filtering by module sends the module substring in the query", async ({
   page,
 }) => {
-  await gotoReady(page, "/logs");
+  await gotoReady(page, LOGS);
 
   await mockLogs(
     page,
@@ -143,7 +146,7 @@ test("filtering by module sends the module substring in the query", async ({
 });
 
 test("paginates to older records and back", async ({ page }) => {
-  await gotoReady(page, "/logs");
+  await gotoReady(page, LOGS);
 
   // Page 0 has a further page; page 1 is the last. Branch on the requested
   // page so Older/Newer render distinct, deterministic rows.
@@ -210,7 +213,7 @@ test("rejects an unauthenticated request to the log endpoint (AC4)", async ({
 });
 
 test("shows an error status when the log read fails", async ({ page }) => {
-  await gotoReady(page, "/logs");
+  await gotoReady(page, LOGS);
 
   await page.route("**/api/rpc/logs", (route) => {
     if (route.request().method() === "POST") {
@@ -235,17 +238,27 @@ test("shows an error status when the log read fails", async ({ page }) => {
   );
 });
 
-test("reaches the log viewer from the settings admin card", async ({
+test("reaches the log viewer from the sidebar Logs item (AC3)", async ({
   page,
 }) => {
   await gotoReady(page, "/settings");
 
-  const card = page.getByTestId("logs-link-card");
-  await expect(card).toBeVisible();
-  await page.getByTestId("logs-link").click();
+  await page.getByTestId("settings-nav-logs").click();
 
-  await expect(page).toHaveURL(/\/logs$/);
+  await expect(page).toHaveURL(/\/settings\?section=logs$/);
   await expect(
-    page.getByRole("heading", { level: 1, name: "Server logs" }),
+    page.getByRole("heading", { level: 2, name: "Server logs" }),
+  ).toBeVisible();
+  await expect(page.getByTestId("logs-page")).toBeVisible();
+});
+
+test("the legacy /logs route redirects into the Logs section", async ({
+  page,
+}) => {
+  await gotoReady(page, "/logs");
+
+  await expect(page).toHaveURL(/\/settings\?section=logs$/);
+  await expect(
+    page.getByRole("heading", { level: 2, name: "Server logs" }),
   ).toBeVisible();
 });
