+++ ui_tests/playwright/tests/flows/settings.spec.ts
@@ -3,6 +3,14 @@ import { expect, test } from "../fixtures/test";
 import { expectMutation } from "../utils/api";
 import { expectNavVisible, gotoReady } from "../utils/nav";
 
+// The Settings page is a permission-gated sidebar (#1324): each configuration
+// group lives behind a `?section=` slug, so tests navigate straight to the
+// section whose controls they exercise. The seeded dev user is an admin, so
+// every section is present.
+const SETTINGS_LIBRARY = "/settings?section=library";
+const SETTINGS_METADATA = "/settings?section=metadata";
+const SETTINGS_EMAIL = "/settings?section=email";
+
 const ebookInput = (page: Page) => page.getByLabel("Ebook Library Path");
 const audiobookInput = (page: Page) =>
   page.getByLabel("Audiobook Library Path");
@@ -41,12 +49,31 @@ const googleBooksStatus = (page: Page) =>
 const googleBooksMessage = (page: Page) =>
   page.getByTestId("google-books-key-status");
 
-test("renders the settings page layout", async ({ page }) => {
-  await page.goto("/settings");
+test("renders the settings sidebar with every admin section", async ({
+  page,
+}) => {
+  await gotoReady(page, "/settings");
 
+  await expect(page.getByTestId("settings-layout")).toBeVisible();
   await expect(
     page.getByRole("heading", { level: 1, name: "Settings" }),
   ).toBeVisible();
+  // The seeded admin sees the Account item plus every server-config section.
+  await expect(page.getByTestId("settings-nav-account")).toBeVisible();
+  await expect(page.getByTestId("settings-nav-library")).toBeVisible();
+  await expect(page.getByTestId("settings-nav-metadata")).toBeVisible();
+  await expect(page.getByTestId("settings-nav-email")).toBeVisible();
+  await expect(page.getByTestId("settings-nav-logs")).toBeVisible();
+  await expectNavVisible(page);
+});
+
+test("renders the Library Location section controls", async ({ page }) => {
+  await gotoReady(page, SETTINGS_LIBRARY);
+
+  await expect(page.getByTestId("settings-nav-library")).toHaveAttribute(
+    "aria-current",
+    "page",
+  );
   await expect(ebookInput(page)).toBeVisible();
   await expect(audiobookInput(page)).toBeVisible();
   await expect(scanIntervalInput(page)).toBeVisible();
@@ -55,19 +82,29 @@ test("renders the settings page layout", async ({ page }) => {
   await expect(settingsStatus(page)).toBeAttached();
   await expect(page.getByTestId("ebook-library-summary")).toBeAttached();
   await expect(page.getByTestId("audiobook-library-summary")).toBeAttached();
-  // F3.3 — the Hardcover suggestions key card renders on the settings page.
+});
+
+test("navigates between sections from the sidebar and updates the URL", async ({
+  page,
+}) => {
+  await gotoReady(page, "/settings");
+
+  // Metadata Lookup — the Hardcover + Google Books key cards.
+  await page.getByTestId("settings-nav-metadata").click();
+  await expect(page).toHaveURL(/\/settings\?section=metadata$/);
   await expect(page.getByTestId("hardcover-key-card")).toBeVisible();
-  // The Google Books metadata-lookup key card renders too.
   await expect(page.getByTestId("google-books-key-card")).toBeVisible();
-  // F4.3 — the admin SMTP (Send-to-Kindle) config card renders too.
+
+  // Email delivery — the SMTP config card.
+  await page.getByTestId("settings-nav-email").click();
+  await expect(page).toHaveURL(/\/settings\?section=email$/);
   await expect(page.getByTestId("smtp-card")).toBeVisible();
-  await expectNavVisible(page);
 });
 
 test("saves the SMTP config and shows a configured status", async ({
   page,
 }) => {
-  await gotoReady(page, "/settings");
+  await gotoReady(page, SETTINGS_EMAIL);
 
   await page.getByLabel("SMTP Host").fill("smtp.example.com");
   await page.getByLabel("From Email").fill("library@example.com");
@@ -109,7 +146,7 @@ test("saves the SMTP config and shows a configured status", async ({
 });
 
 test("saves library paths and shows a success status", async ({ page }) => {
-  await gotoReady(page, "/settings");
+  await gotoReady(page, SETTINGS_LIBRARY);
 
   const ebookPath = "/tmp/omnibus-test-ebooks";
   const audiobookPath = "/tmp/omnibus-test-audiobooks";
@@ -140,7 +177,7 @@ test("saves library paths and shows a success status", async ({ page }) => {
 test("saves an automatic rescan interval alongside the library paths", async ({
   page,
 }) => {
-  await gotoReady(page, "/settings");
+  await gotoReady(page, SETTINGS_LIBRARY);
 
   const ebookPath = "/tmp/omnibus-test-ebooks";
   const audiobookPath = "/tmp/omnibus-test-audiobooks";
@@ -173,7 +210,7 @@ test("saves an automatic rescan interval alongside the library paths", async ({
 test("shows an error status when the server rejects an invalid rescan interval", async ({
   page,
 }) => {
-  await gotoReady(page, "/settings");
+  await gotoReady(page, SETTINGS_LIBRARY);
 
   // Fill every field explicitly (rather than relying on whatever an
   // earlier test left saved) so the expected request body below is
@@ -219,7 +256,7 @@ test("shows an error status when the server rejects an invalid rescan interval",
 });
 
 test("shows an error status when saving settings fails", async ({ page }) => {
-  await gotoReady(page, "/settings");
+  await gotoReady(page, SETTINGS_LIBRARY);
 
   await ebookInput(page).fill("/tmp/whatever");
   await audiobookInput(page).fill("/tmp/whatever-audio");
@@ -262,7 +299,7 @@ test("shows an error status when saving settings fails", async ({ page }) => {
 test("shows an error status when a manual library scan fails", async ({
   page,
 }) => {
-  await gotoReady(page, "/settings");
+  await gotoReady(page, SETTINGS_LIBRARY);
 
   await page.route("**/api/rpc/scan-library", (route) => {
     if (route.request().method() === "POST") {
@@ -294,7 +331,7 @@ test("shows an error status when a manual library scan fails", async ({
 test("saves a Hardcover key, shows a connected status, then clears it", async ({
   page,
 }) => {
-  await gotoReady(page, "/settings");
+  await gotoReady(page, SETTINGS_METADATA);
 
   const key = `hc_test_${Date.now()}`;
 
@@ -337,7 +374,7 @@ test("saves a Hardcover key, shows a connected status, then clears it", async ({
 test("shows an error status when saving the Hardcover key fails", async ({
   page,
 }) => {
-  await gotoReady(page, "/settings");
+  await gotoReady(page, SETTINGS_METADATA);
 
   // Only fail the write — the mount-time GET status fetch must still pass
   // through, otherwise the card never renders the fields under test.
@@ -377,7 +414,7 @@ test("shows an error status when saving the Hardcover key fails", async ({
 test("saves a Google Books key, shows a connected status, then clears it", async ({
   page,
 }) => {
-  await gotoReady(page, "/settings");
+  await gotoReady(page, SETTINGS_METADATA);
 
   const key = `AIza_test_${Date.now()}`;
 
@@ -422,7 +459,7 @@ test("saves a Google Books key, shows a connected status, then clears it", async
 test("shows an error status when saving the Google Books key fails", async ({
   page,
 }) => {
-  await gotoReady(page, "/settings");
+  await gotoReady(page, SETTINGS_METADATA);
 
   // Only fail the write — the mount-time GET status fetch must still pass
   // through, otherwise the card never renders the fields under test.
