+++ ui_tests/playwright/tests/flows/suggestions.spec.ts
@@ -103,10 +103,14 @@ test("shows a connect prompt when no Hardcover key is configured", async ({
       "Add a Hardcover API key to pull read-alikes for this book.",
     ),
   ).toBeVisible();
-  // The seeded dev user is an admin, so the connect card links to Settings.
+  // The seeded dev user is an admin, so the connect card deep-links to the
+  // Metadata Lookup settings section where the Hardcover key lives (#1324).
   const connectLink = page.getByTestId("suggestions-connect-link");
   await expect(connectLink).toBeVisible();
-  await expect(connectLink).toHaveAttribute("href", "/settings");
+  await expect(connectLink).toHaveAttribute(
+    "href",
+    "/settings?section=metadata",
+  );
 });
 
 // ---------------------------------------------------------------------------
