+++ ui_tests/playwright/tests/flows/worker-status.spec.ts
@@ -1,6 +1,6 @@
 // Worker status indicator (issue #69). The indicator polls
 // `/api/rpc/worker_status` at 1 Hz and renders one row per active or
-// recently-finished background task above the Save button on /settings.
+// recently-finished background task above the Save button in the Library Location settings section.
 //
 // We intercept the polling endpoint instead of seeding a real library so
 // the spec doesn't depend on scan duration (which varies wildly with
@@ -56,7 +56,7 @@ test("worker-status indicator stays hidden when the worker is idle", async ({
   page,
 }) => {
   await mockWorkerStatus(page, { active: [], recent_complete: [] });
-  await gotoReady(page, "/settings");
+  await gotoReady(page, "/settings?section=library");
   // The indicator renders nothing — no element with the testid is in the DOM.
   await expect(indicator(page)).toHaveCount(0);
 });
@@ -75,7 +75,7 @@ test("worker-status indicator surfaces an active scan", async ({ page }) => {
     ],
     recent_complete: [],
   });
-  await gotoReady(page, "/settings");
+  await gotoReady(page, "/settings?section=library");
 
   await expect(indicator(page)).toBeVisible();
   await expect(indicator(page)).toContainText("Scanning library");
@@ -96,7 +96,7 @@ test("worker-status indicator shows processed/total when total is known", async 
     ],
     recent_complete: [],
   });
-  await gotoReady(page, "/settings");
+  await gotoReady(page, "/settings?section=library");
 
   await expect(indicator(page)).toContainText("Scanning library");
   await expect(indicator(page)).toContainText("(3 / 10)");
@@ -117,7 +117,7 @@ test("worker-status indicator transitions from active to done and lets the user 
     ],
     recent_complete: [],
   });
-  await gotoReady(page, "/settings");
+  await gotoReady(page, "/settings?section=library");
   await expect(indicator(page)).toContainText("Scanning library");
 
   // Flip the canned response to a terminal Done. Within ~1 polling
@@ -159,7 +159,7 @@ test("worker-status indicator surfaces a failed scan with the error message", as
       },
     ],
   });
-  await gotoReady(page, "/settings");
+  await gotoReady(page, "/settings?section=library");
 
   const failed = page.getByRole("alert");
   await expect(failed).toBeVisible();
